src/requirements.txt
elasticsearch
pyYaml

